ffbs-mesh-vpn-parker: Drop legacy route check
This check was intended to check if the IPv6 address/routing
configuration is sane.
But it currently relies on a hard-coded value for ffbs, so it does not
make sense to have it for all communities.

Lately we did not see this diagnosis trigger, so it's probably safe to
remove anyway.

If we encounter problems later on we can still add a generic check.